requirements.txt
#Data manipulation and analysis
pandas-2.2.2
# Numerical operations
numpy-2.1.0
# Data visualization
matplotlib
seaborn-0.13.2
scipy-1.14.1
# Jupyter notebooks
jupyter
# Web application for interactive dashboards
streamlit-1.37.1
pytest-8.3.2
